requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
PyJWT==2.10.1
python-multipart==0.0.20
jinja2==3.1.5
